requirements.txt
annotated-types==0.7.0
argcomplete==2.0.0
attrs==23.2.0
aws-sam-translator==1.89.0
boto3==1.34.114
botocore==1.34.114
Brotli==1.0.9
cachetools==5.5.2
certifi==2022.9.24
cfn-lint==0.87.4
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
dateparser==1.2.1
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==2.0.1
httplib2==0.22.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jschema-to-python==1.2.3
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
junit-xml==1.9
locust==2.13.0
MarkupSafe==2.1.1
mpmath==1.3.0
msgpack==1.0.4
networkx==3.2.1
packaging==23.0
pbr==6.0.0
pipenv==2023.2.4
pipx==1.1.0
platformdirs==2.6.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==5.9.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.7.2
pydantic_core==2.18.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.28.1
roundrobin==0.0.4
rpds-py==0.18.1
rsa==4.9.1
s3transfer==0.10.1
sarif-om==1.0.4
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sympy==1.12
typing_extensions==4.12.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.25.11
userpath==1.8.0
virtualenv==20.18.0
virtualenv-clone==0.5.7
Werkzeug==2.2.2
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.2
